requirements.txt
# --- Core App ---
streamlit>=1.32.0
streamlit-lottie>=0.0.5
python-dotenv>=1.0.1
requests>=2.31.0

# --- LangChain (Modern 1.x stack) ---
langchain>=1.2.0
langchain-core>=0.2.40
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
langchain-google-genai>=1.0.10

# --- Vector Store ---
faiss-cpu>=1.8.0

# --- Document Loaders ---
pypdf>=4.0.0
python-docx>=1.1.0